Imrpoved basic timing and race handling, Removed LCD code for now. Added printf.h and global.h/cpp

+++ BatterySensor.h
@@ -16,7 +16,7 @@ protected:
 public:
    void init(int iBatterySensorPin);
    void CheckBatteryVoltage();
-   float GetBatteryVoltage();
+   uint16_t GetBatteryVoltage();
    
 private:
    int _iBatterySensorPin;
